requirement.txt
request
selenium
webdriver-manager
pandas
jupyter
